+++ src/Widgets/Forms/InputGroup.php
@@ -22,20 +22,28 @@ class InputGroup extends Widget
     protected function renderContent(): string|Stringable
     {
         $div = Div::make()
+            ->attributes($this->attributes)
             ->addClass('input-group');
 
-        if (null !== $this->prepend) {
-            $div->content(($this->prepend instanceof Div ? $this->prepend : (Div::make()->content($this->prepend)))
-                ->addClass('input-group-prepend'));
+        foreach ($this->prepend as $content) {
+            $div->addContent($this->getAddon($content, 'input-group-prepend'));
         }
 
         $div->addContent(...$this->content);
 
-        if (null !== $this->append) {
-            $div->addContent(($this->append instanceof Div ? $this->append : (Div::make()->content($this->append)))
-                ->addClass('input-group-append'));
+        foreach ($this->append as $content) {
+            $div->addContent($this->getAddon($content, 'input-group-append'));
         }
 
         return $div;
+    }
+
+    /**
+     * Each one is a cell of its own, so two buttons appended read as two, divided the way the addon beside the
+     * input is.
+     */
+    protected function getAddon(string|Stringable $content, string $class): Div
+    {
+        return ($content instanceof Div ? $content : Div::make()->content($content))->addClass($class);
     }
 }