+++ src/Widgets/Forms/Traits/InputGroupTrait.php
@@ -8,18 +8,25 @@ use Stringable;
 
 trait InputGroupTrait
 {
-    protected string|Stringable|null $append = null;
-    protected string|Stringable|null $prepend = null;
+    /**
+     * @var list<string|Stringable>
+     */
+    protected array $append = [];
 
-    public function append(string|Stringable|null $content): static
+    /**
+     * @var list<string|Stringable>
+     */
+    protected array $prepend = [];
+
+    public function append(string|Stringable|null ...$content): static
     {
-        $this->append = $content;
+        $this->append = array_values(array_filter($content));
         return $this;
     }
 
-    public function prepend(string|Stringable|null $content): static
+    public function prepend(string|Stringable|null ...$content): static
     {
-        $this->prepend = $content;
+        $this->prepend = array_values(array_filter($content));
         return $this;
     }
 }